[RHPAM-1179] Support the use of database connection URL for EJB timers
Signed-off-by: Ruben Romero Montes <rromerom@redhat.com>

fixing issues and broken tests

Fix for broken datasource tests.
The jndi name for xa-datasources is not correctly set when using DATASOURCES env when prefix_JDNI is not set, the default value for it would be java:jboss/datasources/prefix and ejb_timer: "<xa-datasource jndi-name="_EJBTimer""
Postgresql with custom drivers are not correctly configured if no PREFIX_URL is set
MySQL pinGlobalTxToPhysicalConnection=true was not being set if PREFIX_URL or XA_PROPERTY_URL is set.
Prepare for MariaDB.