Correctly verify template parameters not used. (#166)
The verification was using the 'Contains' function to verify if a parameter is used or not in the template.
This approach lead to a false positive when the parameter is very similar with the value used somewhere in the template.

i.e. Parameter name "MQ_QUEUES" and similar value used on the DeploymentCOnfig envs: "AMQ_QUEUES" will return no issues.
Which is wrong. This commit fixes this behavior.

Signed-off-by: Filippe Spolti <fspolti@redhat.com>